Added testapp, improved form, requirements.txt updated
To make the first step for new users easier, a testapp was added that can be used to understand how DjangoOpenInviter works.
Beside this, the commit updated the form(s):
- clean_email was a problematic position to load the contacts, since in my tests, password was never ready at this point. Changed to is_valid()
- an example for a form with service selection added

Also added the lxml lib to the requirements.txt, since it doesn't seems to be a integrated part of python.